2.1-flask/requirements.txt
bcrypt==4.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Flask-Bcrypt==1.0.1
Flask==2.2.2
greenlet==2.0.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pip==21.3.1
psycopg2-binary==2.9.5
pydantic==1.10.4
python-dotenv==0.21.1
requests==2.28.2
setuptools==60.2.0
SQLAlchemy==1.4.46
typing-extensions==4.4.0
urllib3==1.26.14
Werkzeug==2.2.2
wheel==0.37.1
zipp==3.12.1
